refactor: extract options static messages
